// client/src/components/layoutBody/ILayoutBodyComponentProps.ts
export interface ILayoutBodyComponentProps {
  /**
   * Styles
   */
  classes?: any

  /**
   * render elements
   */
  children?: any

  /**
   * class to apply
   */
  className?: string

  /**
   * component
   */
  component?: any

  /**
   * all height of the component
   */
  fullHeight?: boolean

  /**
   * all width of component or not
   */
  fullWidth?: boolean

  /**
   * margin or not
   */
  margin?: boolean

  /**
   * margin bottom or not
   */
  marginBottom?: boolean

  /**
   * Style apply to component
   */
  style?: any

  /**
   * width of the component
   */
  width?: any

  /**
   * rest props of the component
   */
  other?: any

}